1. 重构broker调度代码 2. 修复重试间隔问题 3. worker队列为0情况下队列数大小等于任务执行并发数 4. splite分配最多100个 5. 解决部分循环依赖问题 6. 修改参数中枚举类型 防止枚举导致的参数转换异常 7. 修改部分api注释

+++ flowjob-api/src/main/java/org/limbo/flowjob/api/console/vo/PlanVO.java
@@ -20,9 +20,9 @@ package org.limbo.flowjob.api.console.vo;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import org.limbo.flowjob.common.constants.PlanType;
-import org.limbo.flowjob.common.constants.ScheduleType;
-import org.limbo.flowjob.common.constants.TriggerType;
+import org.limbo.flowjob.api.constants.PlanType;
+import org.limbo.flowjob.api.constants.ScheduleType;
+import org.limbo.flowjob.api.constants.TriggerType;
 
 import java.time.LocalDateTime;
 
@@ -57,21 +57,21 @@ public class PlanVO {
      * @see PlanType
      */
     @Schema(title = "任务类型")
-    private Byte planType;
+    private Integer planType;
 
     /**
      * 计划作业调度方式
      * @see ScheduleType
      */
     @Schema(title = "调度方式")
-    private Byte scheduleType;
+    private Integer scheduleType;
 
     /**
      * 计划作业触发方式
      * @see TriggerType
      */
     @Schema(title = "触发方式")
-    private Byte triggerType;
+    private Integer triggerType;
 
     /**
      * 从何时开始调度作业
